rename tiles.js to sf-tiles.js
