+++ src/cards/climates/card.js
@@ -2,7 +2,7 @@ import {html, nothing} from 'lit';
 import {isEqual} from 'lodash-es';
 
 import '../../components/sf-radiator.js';
-import '../../components/tiles.js';
+import '../../components/sf-tiles.js';
 import '../../components/toast.js';
 import {ENTITY_KIND_CLIMATE} from '../../helpers/entities/climate_const.js';
 import {House} from '../../helpers/entities/house.js';
