+++ src/cards/lights/card.js
@@ -2,7 +2,7 @@ import {html, nothing} from 'lit';
 import {unsafeHTML} from 'lit/directives/unsafe-html.js';
 import {isEqual} from 'lodash-es';
 
-import '../../components/tiles.js';
+import '../../components/sf-tiles.js';
 import '../../components/toast.js';
 import {House} from '../../helpers/entities/house.js';
 import {
