+++ src/cards/weather/card.js
@@ -2,7 +2,7 @@ import Chart from 'chart.js/auto';
 import {html, nothing} from 'lit';
 import {isEqual} from 'lodash-es';
 
-import '../../components/tiles.js';
+import '../../components/sf-tiles.js';
 import {
   DailyForecast,
   HourlyForecast,
